# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /plant/bundle:
    get:
      operationId: getPlantBundle
      tags: [plant]
      parameters:
        - name: month
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full plant-level production bundle for the month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantBundle"
  /plant/config:
    get:
      operationId: getPlantConfig
      tags: [plant]
      parameters:
        - name: month
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Plant configuration for the month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantConfigData"
    put:
      operationId: updatePlantConfig
      tags: [plant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantConfigUpdate"
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
    patch:
      operationId: patchPlantConfig
      tags: [plant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantConfigUpdate"
      responses:
        "200":
          description: Config patched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
  /plant/source-config:
    put:
      operationId: upsertPlantSourceConfig
      tags: [plant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantSourceConfigUpsert"
      responses:
        "200":
          description: Source config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
  /plant/cache/invalidate:
    post:
      operationId: invalidatePlantCache
      tags: [plant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [month]
              properties:
                month: { type: string }
      responses:
        "200":
          description: Cache cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
  /plant/trend:
    get:
      operationId: getPlantTrend
      tags: [plant]
      parameters:
        - name: months
          in: query
          required: false
          schema: { type: string }
          description: "Optional filter: a number (e.g. '6' = last 6 months) or comma-separated YYYY-MM list"
      responses:
        "200":
          description: Month-over-month plant trend summaries
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PlantTrendSummary"
  /plant/export/pdf:
    get:
      operationId: exportPlantPdf
      tags: [plant]
      parameters:
        - name: month
          in: query
          required: true
          schema: { type: string }
        - name: section
          in: query
          required: false
          schema:
            type: string
            enum: [control-board, velocity, warnings]
            default: control-board
      responses:
        "200":
          description: PDF export of the plant report
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /plant/weekly-summary:
    get:
      operationId: getPlantWeeklySummary
      tags: [plant]
      summary: Weekly release plan vs actuals for a month
      parameters:
        - name: month
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Weekly release summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantWeeklySummary"
  /plant/snapshots/backfill:
    post:
      operationId: backfillPlantMonitoringSnapshot
      tags: [plant]
      summary: Capture the immutable monitoring snapshot for a closed PTMT month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantSnapshotBackfillInput"
      responses:
        "201":
          description: Snapshot captured or already present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantSnapshotBackfillResult"
        "400":
          description: Invalid month input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantSnapshotBackfillError"
        "409":
          description: The month is not closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantSnapshotBackfillError"
        "422":
          description: Required finalized targets or fresh actuals are unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantSnapshotBackfillError"
        "500":
          description: Snapshot capture failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantSnapshotBackfillError"
...
components:
  securitySchemes:
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: "Managed API key with read scope and optional PTMT/Plumbing segment restrictions."
...
  schemas:
...
    OkResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    NeedsReviewItem:
      type: object
      required: [itemCode, colour, category]
      properties:
        itemCode: { type: string }
        colour: { type: string }
        category: { type: string }
...
    PlantCategoryKPIs:
      allOf:
        - $ref: "#/components/schemas/PlantKPIs"
        - type: object
          required: [category, gapPcs]
          properties:
            category: { type: string }
            gapPcs: { type: number }
...
    PlantContext:
      type: object
      required: [month, snapshotDate, workingDays, elapsed, remaining, shiftsPerDay, shiftHours, lifecycle, workingDaysSource, capturedAt, sourceInfo]
      properties:
        month: { type: string }
        snapshotDate: { type: ["string", "null"] }
        workingDays: { type: integer }
        elapsed: { type: integer }
        remaining: { type: integer }
        shiftsPerDay: { type: integer }
        shiftHours: { type: integer }
        lifecycle: { type: string, enum: [future, open, grace, closed] }
        workingDaysSource: { type: string, enum: [configured, derived] }
        capturedAt: { type: ["string", "null"] }
        sourceInfo:
          type: ["object", "null"]
          additionalProperties: true
    PlantBundle:
      type: object
      required:
        - month
        - context
        - plant
        - categories
        - items
        - dailySeries
        - variancePareto
        - mixFlags
        - warnings
        - recommendations
        - needsReview
        - caveats
        - dataAvailable
        - monitoringStatus
        - targetsAvailable
        - actualsAvailable
        - unavailableReason
      properties:
        month: { type: string }
        context:
          $ref: "#/components/schemas/PlantContext"
        plant:
          $ref: "#/components/schemas/PlantKPIs"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PlantCategoryKPIs"
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlantItemKPIs"
        dailySeries:
          type: array
          items:
            $ref: "#/components/schemas/PlantDayRecord"
        variancePareto:
          type: array
          items:
            $ref: "#/components/schemas/PlantItemKPIs"
        mixFlags:
          type: array
          items:
            $ref: "#/components/schemas/PlantMixFlag"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/PlantWarning"
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/PlantRecommendation"
        needsReview:
          type: array
          items:
            $ref: "#/components/schemas/NeedsReviewItem"
        caveats:
          type: array
          items: { type: string }
        dataAvailable: { type: boolean }
        monitoringStatus: { type: string, enum: [live, grace, frozen, unavailable, future] }
        targetsAvailable: { type: boolean }
        actualsAvailable: { type: boolean }
        unavailableReason: { type: ["string", "null"] }
    PlantSnapshotBackfillInput:
      type: object
      required: [month]
      properties:
        month:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
    PlantSnapshotBackfillResult:
      type: object
      required: [month, status, capturedAt]
      properties:
        month: { type: string }
        status: { type: string, enum: [frozen] }
        capturedAt: { type: string }
    PlantSnapshotBackfillError:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
        month: { type: string }
    PlantTrendSummary:
      type: object
      required: [month, workingDays]
      properties:
        month: { type: string }
        workingDays: { type: integer }
        attainmentMaxPct: { type: ["number", "null"] }
        attainmentMinPct: { type: ["number", "null"] }
        avgDailyPcs: { type: ["number", "null"] }
        linearityIndex: { type: ["number", "null"] }
        producedTotal: { type: ["number", "null"] }
        targetMax: { type: ["number", "null"] }
        targetMin: { type: ["number", "null"] }
        bestCategory: { type: ["string", "null"] }
        worstCategory: { type: ["string", "null"] }
        ragBand: { type: ["string", "null"] }
...
    PlantConfigData:
      type: object
      required: [month, workingDays, workingDaysSource, shiftsPerDay, shiftHours, snapshotDate, sourceConfigs]
      properties:
        month: { type: string }
        workingDays: { type: integer }
        workingDaysSource: { type: string, enum: [configured, derived] }
        shiftsPerDay: { type: integer }
        shiftHours: { type: integer }
        snapshotDate: { type: ["string", "null"] }
        sourceConfigs:
          type: array
          items:
            $ref: "#/components/schemas/PlantSourceConfigItem"
    PlantConfigUpdate:
      type: object
      required: [month]
      properties:
        month: { type: string }
        workingDays: { type: [integer, "null"] }
        shiftsPerDay: { type: integer }
        shiftHours: { type: integer }
        snapshotDate: { type: ["string", "null"] }
        thresholds: { type: object, additionalProperties: true }
...
    PlantSourceConfigUpsert:
      type: object
      required: [month, fileId]
      properties:
        month: { type: string }
        fileId: { type: string }
        notes: { type: string }
...
    PlantWeeklySummary:
      type: object
      required: [month, snapshotDate, weekCalendar, currentWeek, elapsedDaysInWeek, plant, categories]
      properties:
        month: { type: string }
        snapshotDate:
          nullable: true
          type: string
        weekCalendar:
          type: array
          items:
            $ref: "#/components/schemas/PlantWeekWindow"
        currentWeek: { type: integer }
        elapsedDaysInWeek: { type: integer }
        plant:
          $ref: "#/components/schemas/PlantWeeklyPlantSummary"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PlantWeeklyCategoryRow"
...